make sentinel compatible with local socket servers
The server list will keep track of local socket servers and regular
TCP servers. The second element of each list inside of the server list
will be a port number for normal TCP servers and a path-like value for
local socket servers.

The sentinel messages are working in a similar way, the messages
will just get passed a path-like value for local socket servers or a
port number for regular TCP servers but
otherwise not know the difference. This should allow old code to keep
functioning like normal while also allowing us to reuse the logic for
local socket servers.

In the case of stop-server we can provide a key which associates to a
list element when searching the list of servers.
Passing the key :port will actually work on local socket servers,
but an extra :path case has been added to allow users to call
stop-server on local socket servers using :path instead of :port for
clarity purposes.

example: (send-to-sentinel '(:stop-server :path ,local-socket-path))